fix being able to kill daemon on linux

+++ src/daemon.rs
@@ -2,7 +2,10 @@
 use std::fs;
 
 #[cfg(target_os = "linux")]
-use ::{anyhow::Context, libc::SIGINT};
+use ::{
+    anyhow::Context,
+    libc::{SIGINT, geteuid},
+};
 use anyhow::Result;
 use iter_tools::Itertools;
 
@@ -48,33 +51,42 @@ pub fn kill() -> Result<()> {
 
 #[cfg(target_os = "linux")]
 pub fn kill() -> Result<()> {
-    fw::stop()?;
+    if unsafe { geteuid() == 0 } {
+        fw::stop()?;
 
-    let my_exe_path = fs::read_link("/proc/self/exe")?;
+        let my_exe_path = fs::read_link("/proc/self/exe")?;
 
-    let procs_readdir = std::fs::read_dir("/proc")?;
-    for proc in procs_readdir {
-        let proc = proc?;
+        let procs_readdir = std::fs::read_dir("/proc")?;
+        for proc in procs_readdir {
+            let proc = proc?;
 
-        let Ok(proc_pid) = i32::from_str_radix(&proc.file_name().to_string_lossy(), 10) else {
-            continue;
-        };
+            let Ok(proc_pid) = i32::from_str_radix(&proc.file_name().to_string_lossy(), 10) else {
+                continue;
+            };
 
-        let Ok(cmd) = std::fs::read_to_string(proc.path().join("cmd")) else {
-            continue;
-        };
-        if !cmd.contains("--daemon") {
-            continue;
+            let Ok(cmd) = std::fs::read_to_string(proc.path().join("cmd")) else {
+                continue;
+            };
+            if !cmd.contains("--daemon") {
+                continue;
+            }
+
+            let exe_path = proc.path().join("exe");
+            let Ok(exe_path) =
+                fs::read_link(&exe_path).context(exe_path.clone().display().to_string())
+            else {
+                continue;
+            };
+            if exe_path == my_exe_path {
+                unsafe { libc::kill(proc_pid, SIGINT) };
+            }
         }
-
-        let exe_path = proc.path().join("exe");
-        let Ok(exe_path) = fs::read_link(&exe_path).context(exe_path.clone().display().to_string())
-        else {
-            continue;
-        };
-        if exe_path == my_exe_path {
-            unsafe { libc::kill(proc_pid, SIGINT) };
-        }
+    } else {
+        std::process::Command::new("/usr/bin/env")
+            .arg("pkexec")
+            .arg(std::env::current_exe()?)
+            .arg("--kill")
+            .spawn()?;
     }
 
     Ok(())
